requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.4
colorama==0.4.6
exceptiongroup==1.2.2
fastavro==1.10.0
filelock==3.17.0
Flask==3.1.0
fsspec==2024.12.0
googletrans==4.0.2
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
libretranslatepy==2.1.1
lxml==5.3.0
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
outcome==1.3.0.post0
packaging==24.2
parameterized==0.9.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.2
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
tokenizers==0.21.0
tqdm==4.67.1
translate==3.6.1
trio==0.28.0
trio-websocket==0.11.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
pydub==0.25.1
googlesearch-python==1.3.0